fix: resolve all test compilation errors and unused imports
This commit fixes compilation errors across the test suite:

Test Compilation Fixes:
- Fixed duplicate pool package imports in multiple test files
- Updated pool.GlobalBufferPool.Get() calls to include size parameter
- Replaced []byte with bytes.Buffer for writer operations
- Updated Record types from models.Record to pool.Record where appropriate
- Fixed field type constants (FieldTypeInteger to FieldTypeInt, etc.)
- Updated configuration usage to config.BaseConfig

Import and Variable Fixes:
- Removed unused imports in benchmark tests
- Fixed unused variable declarations
- Added missing helper functions (createTestData, createTestCSV, removeTestFile)
- Updated transform function signatures to match pipeline interfaces

Files Fixed:
- internal/pipeline/streaming_benchmark_test.go
- pkg/compression/compression_bench_test.go
- pkg/connector/destinations/bigquery/bigquery_destination_test.go
- pkg/formats/columnar/columnar_bench_test.go
- pkg/json/json_pool_test.go
- pkg/performance/performance_bench_test.go
- tests/benchmarks/*.go (multiple benchmark test files)
- tests/integration/snowflake_optimized_test.go

All test files now compile successfully. Remaining linter warnings are
code quality issues (duplication, unused parameters, whitespace) not
blocking compilation.

+++ tests/benchmarks/hybrid_storage_test.go
@@ -300,7 +300,7 @@ func BenchmarkDirectStorageAdapter(b *testing.B) {
 					record.SetData("id", fmt.Sprintf("%d", j))
 					record.SetData("value", fmt.Sprintf("value_%d", j))
 					record.SetData("category", fmt.Sprintf("cat_%d", j%10))
-					record.SetData("amount", j*1.5)
+					record.SetData("amount", float64(j)*1.5)
 					
 					adapter.AddRecord(record)
 					record.Release()
